+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/monitor/sampling/MetricFetcher.java
@@ -10,7 +10,7 @@ import com.linkedin.cruisecontrol.metricdef.MetricDef;
 import com.linkedin.kafka.cruisecontrol.exception.SamplingException;
 import com.linkedin.kafka.cruisecontrol.monitor.sampling.holder.BrokerMetricSample;
 import com.linkedin.kafka.cruisecontrol.monitor.sampling.holder.PartitionMetricSample;
-import java.util.HashSet;
+import java.util.Collections;
 import java.util.Set;
 import java.util.concurrent.Callable;
 import org.apache.kafka.common.Cluster;
@@ -59,15 +59,16 @@ abstract class MetricFetcher implements Callable<Boolean> {
                 Timer fetchTimer,
                 Meter fetchFailureRate,
                 MetricSampler.SamplingMode samplingMode) {
-    this(metricSampler, cluster, sampleStore, null, assignedPartitions, startTimeMs,
-        endTimeMs, metricDef, fetchTimer, fetchFailureRate, samplingMode);
+    this(metricSampler, cluster, sampleStore, null, assignedPartitions, startTimeMs, endTimeMs,
+         metricDef, fetchTimer, fetchFailureRate, samplingMode);
   }
 
   /**
    * @param metricSampler The sampler used to retrieve metrics.
    * @param cluster The Kafka cluster.
    * @param sampleStore Sample store to persist the fetched samples, or skip storing samples if {@code null}.
-   * @param sampleStoreForPartitionMetricsDuringExecution Sample store to persist the fetched partition samples during execution
+   * @param sampleStoreForPartitionMetricsDuringExecution Sample store to persist the fetched partition samples during execution, or
+   *                                                      skip storing samples if {@code null}.
    * @param assignedPartitions Partitions to fetch samples from.
    * @param startTimeMs The start time of the sampling period.
    * @param endTimeMs The end time of the sampling period.
@@ -77,16 +78,16 @@ abstract class MetricFetcher implements Callable<Boolean> {
    * @param samplingMode The mode of sampling to indicate the sample type of interest.
    */
   MetricFetcher(MetricSampler metricSampler,
-      Cluster cluster,
-      SampleStore sampleStore,
-      SampleStore sampleStoreForPartitionMetricsDuringExecution,
-      Set<TopicPartition> assignedPartitions,
-      long startTimeMs,
-      long endTimeMs,
-      MetricDef metricDef,
-      Timer fetchTimer,
-      Meter fetchFailureRate,
-      MetricSampler.SamplingMode samplingMode) {
+                Cluster cluster,
+                SampleStore sampleStore,
+                SampleStore sampleStoreForPartitionMetricsDuringExecution,
+                Set<TopicPartition> assignedPartitions,
+                long startTimeMs,
+                long endTimeMs,
+                MetricDef metricDef,
+                Timer fetchTimer,
+                Meter fetchFailureRate,
+                MetricSampler.SamplingMode samplingMode) {
     _metricSampler = metricSampler;
     _cluster = cluster;
     _sampleStore = sampleStore;
@@ -128,17 +129,16 @@ abstract class MetricFetcher implements Callable<Boolean> {
       MetricSampler.Samples samples = fetchSamples();
       if (_sampleStore != null) {
         if (_samplingMode == MetricSampler.SamplingMode.ONGOING_EXECUTION) {
-          _sampleStore.storeSamples(new MetricSampler.Samples(new HashSet<>(), samples.brokerMetricSamples()));
+          _sampleStore.storeSamples(new MetricSampler.Samples(Collections.emptySet(), samples.brokerMetricSamples()));
         } else {
           _sampleStore.storeSamples(samples);
         }
       }
 
-      //replace _sampleStore
-      if (_samplingMode == MetricSampler.SamplingMode.ONGOING_EXECUTION && _sampleStoreForPartitionMetricsDuringExecution
-          != null) {
+      if (_samplingMode == MetricSampler.SamplingMode.ONGOING_EXECUTION &&
+          _sampleStoreForPartitionMetricsDuringExecution != null) {
         _sampleStoreForPartitionMetricsDuringExecution.storeSamples(new MetricSampler.Samples(samples.partitionMetricSamples(),
-            new HashSet<>()));
+                                                                                              Collections.emptySet()));
       }
       // TODO: evolve sample store interface to allow independent eviction time for different type of metric samples.
       // We are not calling sampleStore.evictSamplesBefore() because the broker metric samples and partition metric
@@ -166,9 +166,7 @@ abstract class MetricFetcher implements Callable<Boolean> {
     if (_samplingMode == MetricSampler.SamplingMode.ALL || _samplingMode == MetricSampler.SamplingMode.PARTITION_METRICS_ONLY) {
       usePartitionMetricSamples(samples.partitionMetricSamples());
     }
-    if (_samplingMode == MetricSampler.SamplingMode.ALL ||
-        _samplingMode == MetricSampler.SamplingMode.BROKER_METRICS_ONLY ||
-        _samplingMode == MetricSampler.SamplingMode.ONGOING_EXECUTION) {
+    if (_samplingMode != MetricSampler.SamplingMode.PARTITION_METRICS_ONLY) {
       useBrokerMetricSamples(samples.brokerMetricSamples());
     }
