+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/monitor/sampling/MetricFetcherManager.java
@@ -140,6 +140,7 @@ public class MetricFetcherManager {
    * @param endMs the end time of the fetching period.
    * @param timeoutMs the timeout.
    * @param sampleStore the sample store to save the broker and partition metric samples.
+   * @param samplingMode the sampling mode to indicate which type of samples is interested.
    * @return True if there was no fetching error, false otherwise.
    */
   public boolean fetchMetricSamples(long startMs,
@@ -156,14 +157,16 @@ public class MetricFetcherManager {
    * @param endMs the end time of the fetching period.
    * @param timeoutMs the timeout.
    * @param sampleStore the sample store to save the broker and partition metric samples.
+   * @param sampleStoreForPartitionMetricsDuringExecution the sample store to save the partition metrics samples during ongoing execution.
+   * @param samplingMode the sampling mode to indicate which type of samples is interested.
    * @return True if there was no fetching error, false otherwise.
    */
   public boolean fetchMetricSamples(long startMs,
-      long endMs,
-      long timeoutMs,
-      SampleStore sampleStore,
-      SampleStore sampleStoreForPartitionMetricsDuringExecution,
-      MetricSampler.SamplingMode samplingMode) {
+                                    long endMs,
+                                    long timeoutMs,
+                                    SampleStore sampleStore,
+                                    SampleStore sampleStoreForPartitionMetricsDuringExecution,
+                                    MetricSampler.SamplingMode samplingMode) {
     LOG.info("Kicking off metric sampling for time range [{}, {}], duration {} ms with timeout {} ms.",
              startMs, endMs, endMs - startMs, timeoutMs);
     Set<TopicPartition> partitionAssignment = _partitionAssignor.assignPartitions(_metadataClient.cluster());
