+++ docs/wiki/User Guide/Configurations.md
@@ -79,7 +79,8 @@ The following configurations are inherited from the open source Kafka client con
 | leader.network.inbound.weight.for.cpu.util                    | Double  | N         | 0.6                                                                                     | Kafka Cruise Control uses the following model to derive replica level CPU utilization: REPLICA_CPU_UTIL = a * LEADER_BYTES_IN_RATE + b * LEADER_BYTES_OUT_RATE + c * FOLLOWER_BYTES_IN_RATE. This configuration will be used as the weight for LEADER_BYTES_IN_RATE.                                                                                                                                                |
 | leader.network.outbound.weight.for.cpu.util                   | Double  | N         | 0.1                                                                                     | Kafka Cruise Control uses the following model to derive replica level CPU utilization: REPLICA_CPU_UTIL = a * LEADER_BYTES_IN_RATE + b * LEADER_BYTES_OUT_RATE + c * FOLLOWER_BYTES_IN_RATE. This configuration will be used as the weight for LEADER_BYTES_OUT_RATE.                                                                                                                                               |
 | follower.network.inbound.weight.for.cpu.util                  | Double  | N         | 0.3                                                                                     | Kafka Cruise Control uses the following model to derive replica level CPU utilization: REPLICA_CPU_UTIL = a * LEADER_BYTES_IN_RATE + b * LEADER_BYTES_OUT_RATE + c * FOLLOWER_BYTES_IN_RATE. This configuration will be used as the weight for FOLLOWER_BYTES_IN_RATE.                                                                                                                                              |
-| sample.store.class                                            | Class   | N         | com.linkedin.kafka.cruisecontrol.monitor.sampling.KafkaSampleStore                      | The sample store class name. User may configure a sample store that persist the metric samples that have already been aggregated into Kafka Cruise Control. Later on the persisted samples can be reloaded from the sample store to Kafka Cruise Control.                                                                                                                                                           |
+| sample.store.class                                            | Class   | N         | com.linkedin.kafka.cruisecontrol.monitor.sampling.KafkaSampleStore                      | The sample store class name. User may configure a sample store that persists the metric samples that have already been aggregated into Kafka Cruise Control. Later on the persisted samples can be reloaded from the sample store to Kafka Cruise Control.                                                                                                                                                          |
+| sample.partition.metrics.store.during.execution.class         | Class   | N         | com.linkedin.kafka.cruisecontrol.monitor.sampling.NoopSampleStore                       | The sample store during execution class name. User may configure a sample store that persists the partition metric samples collected while there is an ongoing execution. This config is not intended for enhancing the fault tolerance of the system through recovery of the historical load information.                                                                                                          |                         
 | max.allowed.extrapolations.per.partition                      | Integer | N         | 5                                                                                       | The maximum allowed number of extrapolations for each partition. A partition will be considered as invalid if the total number extrapolations in all the windows goes above this number.                                                                                                                                                                                                                            |
 | max.allowed.extrapolations.per.broker                         | Integer | N         | 5                                                                                       | The maximum allowed number of extrapolations for each broker. A broker will be considered as invalid if the total number extrapolations in all the windows goes above this number.                                                                                                                                                                                                                                  |
 | partition.metric.sample.aggregator.completeness.cache.size    | Integer | N         | 5                                                                                       | The metric sample aggregator caches the completeness metadata for fast query. The completeness describes the confidence level of the data in the metric sample aggregator. It is primarily measured by the validity of the metrics samples in different windows. This configuration configures The number of completeness cache slots to maintain.                                                                  |
